// src/constants/enums.ts
export enum Status {
    INACTIVE = 'INACTIVE',
    ACTIVE = 'ACTIVE',
}

export enum StatusRoom {
    CHECKED = 'CHECKED',
    UNCHECKED = 'UNCHECKED',
    HAVE_PROBLEM = 'HAVE_PROBLEM',
}

export enum TypeCompletion {
    COMPLETED = 'COMPLETED',
    UNCOMPLETED = 'UNCOMPLETED',
}

export enum UserType {
    OPERATOR = 'OPERATOR',
    TEACHER = 'TEACHER',
    ADMINISTRATIVE_STAFF = 'ADMINISTRATIVE_STAFF',
    ADMIN = 'ADMIN',
}

export enum TypeMetadata {
    PLACE = 'PLACE',
    FLOOR = 'FLOOR',
}